Merging skeletal muscle of arm into muscle of arm.
The former was a type of muscle tissue, the muscle organ is the more
useful grouping class, and consistent with other ontologies such as
the FMA.

The skeletal vs generic muscle distinction is not particularly useful
for limbs and skeletally defined subdivisions as these will by default
be skeletal muscle organs. We experimentally add a hidden GCI that says
every limb muscle organ is made of skeletal muscle tissue. This would
be violated if we decide that, say, some portion of the smooth muscle
of a vein of your arm was a muscle organ. So we are saying something
quite strong about muscle organs here.