uberon_edit.obo use oboformat v1.4 tag ordering
Reserialize uberon_edit.obo using pyontutils.obo_io so that the tag
ordering in various blocks matches the v1.4 format spec.
https://owlcollab.github.io/oboformat/doc/obo-syntax.html

The primary changes are in the header ordering and in the ordering of
the property_value tag.

One other slight change is that trailing qualifiers are ordered
alphabetically. In theory this helps to ensure deterministic
serialization of qualifiers.

uberon_edit.obo cannot be edited using Protege because Protege
does not currently roundtrip many tag types and also does not properly
roundtrip xrefs, failing to serialize the compact version of the xref as
it was originally provided, so this reserializaztion will have no effect
on any Protege related workflow.