Fixes #1641
This is complicated by the different ways the term has been used; see the discussion in the Invertebrates textbook: https://archive.org/details/invertebratesnde00brus/page/n240/mode/2up (which also talks about vertebrate usage)

This also discusses different lineages. In inverts largely from ectoderm, in verts from mesoderm -- but I think this is vastly simplified. In think in non-vert chordates and even other deuterostomes it comes from mesoderm. And in verts we have classes in uberon such as portions of mesoderm from neural crest... which is ectoderm.

For now I am simplifying definition to be aggregate of mesenchymal cells (see linked ticket for discussion of MC), and making the lineage restriction a has-developmental-contribution-from with taxon-GCI to vert.