Obsolete 'insect anterior-posterior subdivision of organism'.
The term 'insect anterior-posterior subdivision of organism'
(UBERON:6057001) was likely introduced as a port of FBbt's
'anterior-posterior subdivision of organism' (FBbt:00057001). However,
while the term makes sense in FBbt (it groups the several kinds of
subdivisions of the insect organism: segments, parasegments, and
tagmata), it is of dubious use in UBERON, where the taxon-neutral term
'subdivision of organism along main body axis' (UBERON:0011676) has the
same meaning.

It is also used incorrectly, as it is a parent to 'panarthropod tagma'
(UBERON:6000002) (another port from FBbt, generalized to the broader
arthropods) even though it is insect-specific.

Dubious usefulness and incorrect usage are enough reasons to obsolete
it.