Rename 'panarthropod tagma' to 'arthropod tagma'.
Common usage restricts the term 'tagma' to arthropods, and it seems
unclear whether this can be extended to the broader panarthropods
(tardigrades and velvet worms).

Since this is one of only two terms we have in UBERON pertaining to
panarthropods (the other one being 'panarthropod head', which is also of
dubious usefulness), I propose to restrict it strictly to arthropods.

The definition is also updated, as not all arthropods have three
tagmata.