Silence warning about closed file handle.
The make-bridge-ontologies-from-xrefs.pl script attempts to open a
non-existing ftr.obo file and read from it regardless of whether it has
been opened successfully, resulting in a bunch of warnings when
generating the bridge files.

Not sure why the script attempts to read that inexisting file, so I
won't remove the corresponding code, but at least let's handle the
absence of the file gracefully.

closes #2144