Do not generate tmp/seed.obo from tmp/seed.owl.
The tmp/seed.owl file is generated by merging tmp/uberon-edit.owl and
tmp/cl-core.obo, something which is (presumably) needed further down the
line (tmp/seed.owl is used as input in several other rules).

However, to generate the bridge files, we do *not* want to use a file in
which cl-core has been merged, because otherwise we end up with Uberon
bridge files that also contain axioms for CL terms.

So we generate the tmp/seed.obo (whose sole purpose is to generate the
Uberon bridge files) directly from the edit file, and we only make sure
the mappings.owl component (containing the xrefs from which the mappings
are to be derived) is merged.