Obsolete children of `insect anlage in statu nascendi`.
The two terms `clypeo-labral anlage in statu nascendi` and `visual
anlage in statu nascendi` are way too specific to belong in Uberon.
Having the parental term `insect anlage in statu nascendi` (mapped to
the corresponding term in FBbt) should be enough--anyone looking for a
more precise term should go look for the children of the FBbt term.

Obsoleting `clypeo-labral anlage in statu nascendi` allows to remove the
FBbt term `anterior ectoderm anlage`, which was imported into Uberon
since `clypeo-labral anlage in statu nascendi` has a part_of
relationship to it. (As stated in #2025, we do not want any Uberon term
to have any kind of relationship to a species-specific foreign
ontology.)

Obsoleting these two terms has cascading consequences:

* `insect clypeo-labral anlage` is the starting point of a chain of
  `develops_from` relationships that goes up to `insect clypeus`;
* similarly, `insect visual anlage in statu nascendi` is the starting
  point of a chain that goes up to `insect Bolwig organ`.

It is doubtful that all these terms are really useful, but I don’t want
to obsolete them all for now, so this PR simply breaks the chains at
their beginnings by removing the first `develops_from` axioms. Deciding
what to do with all the "insect anlage" and "insect primordium" terms
will need to be done later (probably as part of a larger review of all
the 6xxxxxx terms).

close #2154