Automatically ping PRs after one month of inactivity.
Update the 'stale-issues' workflow to implement the decision mentioned
in #2268.

Any pull request that has not seen any activity in one month will be
tagged with `unreviewed` and a comment will be added with the following
message:

"This PR has not seen any activity in the past month; if nobody comments
or reviews it in the next week, the PR editor will be allowed to proceed
with merging without explicit approval, should they wish to do so."

Pull requests may be explicitly marked with a tag `approval required` to
prevent them from receiving this comment. This way, a PR editor (or
anybody else) can express that they do wish the PR to be explicitly
approved.

closes: #2268