Fix `egg chamber`.
The term `egg chamber` (UBERON:0003199), intended to refer to the insect
egg chamber (FBbt:00004894) had no definition and an incorrect
classification.

In insects, the egg chamber is *not* an anatomical space where the egg
would be, but is instead the follicle, comprising the oocyte, the
accompanying nurse cells, and a surrounding layer of follicle cells.

We update the classification accordingly, add a proper definition, and a
taxon constraint to restrict the term to insects.

We also remove the mapping with HAO:0000287 (hymenoptera `anellus`)
which seems completely bogus.

closes #476