Polysemy: gasserian ganglion. Delete offending UBERON:3011045
This PR tries to resolve the 1 of 4 of issues mentioned in #2424

The problem that is addressed here is that

"gasserian ganglion" (UBERON:3011045) is an exact synonym to "trigeminal ganglion" (UBERON:0001675) due to WIKIPEDIA. This creates unecessary polysemy.

The proposed solution is to eliminate UBERON:3011045 altogether. There is a comment there that reads
> property_value: provenance_notes "This class was sourced from an external ontology (amphibian_anatomy). Its definitions, naming conventions and relationships may need to be checked for compatibility with uberon" xsd:string {source="http://purl.obolibrary.org/obo/aao.owl"}

I am not sure I follow proper procedure here so if there is any feedback it much appreciated.

As requested by @matentzn I am tagging him.