Polysemy: gasserian ganglion. Obsolete offending UBERON:3011045 (try: 3)
This PR tries to resolve the 1 of 4 of issues mentioned in obophenotype#2424

The problem that is addressed here is that

"gasserian ganglion" (UBERON:3011045) is an exact synonym to "trigeminal ganglion" (UBERON:0001675) per Wikipedia. This creates unecessary polysemy.

The proposed solution is to obsolete UBERON:3011045 altogether. Theare 2 points that strengthen this decision
1) Existing provenance reads as "This class was sourced from an external ontology (amphibian_anatomy). Its definitions, naming conventions and relationships may need to be checked for compatibility with uberon"
2) In the obsoleted term there was a synonym (gasserian ganglia) that was already listed and tagged as plural in the UBERON:0001675