bridges: Change relation to life cycle terms to BFO:0000066.
Update the bridge generation scripts so that, in species-specific
bridges, terms that are cross-referenced with Uberon's `life cycle` (or
its subclasses) are defined in the generated bridges as equivalent to
the Uberon term 'occuring in' (BFO:0000066) the specific taxon, rather
than the Uberon term 'being part of' (BF:0000050) the specific taxon.

This is dependent on having a list of all terms that are
cross-referenced to the Uberon's `life cycle` hierarchy. Generating such
a list is outside the scope of this commit.