Polysemy: temporal pole. Change in UBERON:0006479 to RELATED
This PR tries to resolve the 2 of 4 of issues mentioned in #2424

The problem that is addressed here is that

"temporal pole" in UBERON:0006479 is listed as an EXACT synonym of "Brodmann (1909) area 38". However, "temporal pole" is also the canocical for UBERON:0002576. This creates unecessary polysemy.

Looks like UBERON:0006479 is a more specific term (area 38). Therefore the proposed solution changes EXACT to RELATED in "temporal pole".