Polysemy: trochlea of femur. Obsolete offending UBERON:7500077
This PR tries to resolve the 3 of 4 of the issues mentioned in #2424

The problem that is addressed here is that

"trochlea of femur" (UBERON:7500077) is an exact synonym to "intercondylar fossa" (UBERON:4200096). This creates unecessary polysemy.

The proposed solution is to obsolete UBERON:7500077 altogether.

I also moved from the obsoleted term the following:
* synonym: "patellar surface of femur" EXACT []
* is_a: UBERON:4200230 ! surface of bone
* relationship: part_of UBERON:0004406 ! distal epiphysis of femur

Tagged @chandron since he helped with the biology.