Polysemy: PHY. Demote synonymity to RELATED in UBERON:8440028
This PR tries to resolve the 6 of 6 of the issues mentioned in #2424 (There are 2 new ones from the last time check was run)

The issue also blocks QC PR #2640

The problem that is addressed here is that:

* "phy" is listed as an EXACT synonym of UBERON:2000438 with name: "parhypural"
* "PHY" is listed as an EXACT synonym of UBERON:8440028 with name: "Perihypoglossal nuclei"

The proposed solution demotes PHY to RELATED for "Perihypoglossal nuclei".

In my opinion both synonyms should be deleted since they represent a significant risk of misinterpretation by any
automated system. Having said the path of minimum change was selected for this PR.

Hopefully this PR is the last in the polysemy PR series.

Tagged @elekka since she helped with the biology and @matentzn as requested.