Fix classification of `imaginal disc`.
`imaginal disc` is classified as `epithelium sac`, which ultimately
makes it a single `epithelium`. This is incorrect as an imaginal disc is
made of two distinct epithelia: the “disc proper” (a columnar
epithelium) and the “peripodial membrane” (a squamous epithelium). It
should therefore be classified as a `multi-tissue structure` rather than
as a single epithelium.

closes #2684