Fix classification of `non-connected functional system`.
`non-connected functional system` is classified as a `anatomical
system`. This is wrong as `anatomical system` is explicitly defined as a
_connected_ system (a system made of parts that are all connected). With
such a definition a non-connected functional system obviously cannot be
an anatomical system. Instead, it should be a `disconnected anatomical
group`, which is specifically intended for entities made of parts that
are not all connected to each other.

Likewise for `insect non-connected developing system`, for the same
reason.