Fix classification of `anatomical cluster`.
`anatomical cluster` is simply classified as `anatomical entity`.
However, it is defined as an “anatomical group that has its parts
adjacents to one another”. This does not seem to imply that the
component parts are necessarily all _connected_ to each other, so it can
be classified more precisely as `disconnected anatomical group` instead.