Fix classification of `insect synaptic neuropil block`.
`insect synaptic neuropil block` is defined as a “closely associated
group of synaptic neuropil domains”, indicating it is a _disconnected_
group (“closely associated” does not imply “connected”). However it is
classified as `anatomical group`, which (as a subclass of `anatomical
structure`) makes it a connected structure.

We re-classify instead as a `disconnected anatomical group`. This
mirrors the classification of the equivalent class in FBbt.