Fix classification of `insect synaptic neuropil`.
`insect synaptic neuropil` is classified as `anatomical group`, which
makes it a connected anatomical structure. This is too strong an
assertion. In FBbt, the equivalent class `synaptic neuropil` is
classified simply as an “anatomical entity”, which makes no statement
about the “connectedness” of the entity, and we have subclasses for both
connected and non-connected neuropils).

We re-classify `insect synaptic neuropil` as a `material anatomical
entity` to mirror FBbt’s classification and avoid implying that neuropil
are necessarily connected structures.