Fix classification of non-connected systems.
`endocrine system` is classified as `anatomical system`, which makes it
a _connected system_ (because `anatomical system` is a `anatomical
structure`, which is explicitly defined as refering to a structure whose
components are all connected to each other). But the various glands that
make up the endocrine system are _not_ connected to each other, so that
system should be a `non-connected functional system` instead.

Likewise for `immune system`, whose components (e.g. spleen, thymus,
immune cells, etc.) are spread throughout the body. (It could be argued
that they are ultimately connected by blood or lymphatic vessels, but I
believe this pushing the defintion of “connected” too far.)

Likewise for `sensory system`. Again, it could be argued that the sense
organs that make up the sensory system are ultimately all connected by
nerves, but by that logic almost everything in the body can be said to
be “connected”.