Check for both Label and Synonym polysemy across Terms (#2681)
* Check for both Label and Synonym polysemy across Terms

This PR detects synonym-synonym and label-synonym across different Terms (UBERON Ids).
Another implementation (but with the same logic and functionality) of this check was
used to detect and fix (merged) the followings issues:

* https://github.com/obophenotype/uberon/pull/2465
* https://github.com/obophenotype/uberon/pull/2470
* https://github.com/obophenotype/uberon/pull/2506

The backstory is in https://github.com/obophenotype/uberon/issues/2424

The existing check
https://robot.obolibrary.org/report_queries/duplicate_exact_synonym detects
only synonym-synonym violations. It was reported in
https://github.com/ontodev/robot/issues/607 but closed at the time.

The existing check
https://robot.obolibrary.org/report_queries/duplicate_label_synonym
detects all label-synonym duplication both in the same Term and across
different terms. The duplicity in the same term is a warning but the
duplicity across different terms should be an error.

In this PR the test is case insensitive because downstream applications
have a strong possibility to be case insensitive in order to maximize
coverage.

Finally I have added an exemption for ABBREVIATION per @shawntanzk suggestion for
cases like #2643, #2641.

The errors should look like this. Note that for the example below a fix is already
merged and thus it is not a problem anymore.

entity,property,value
http://purl.obolibrary.org/obo/UBERON_0002633,http://www.geneontology.org/formats/oboInOwl#hasExactSynonym,nV
http://purl.obolibrary.org/obo/UBERON_2005093,http://www.geneontology.org/formats/oboInOwl#hasExactSynonym,NV

Thanks @matentzn and @anitacaron for the guidance and the hand-holding.

--
Vassilis Virvilis

* small fix

Co-authored-by: Anita Caron <anitacaron@users.noreply.github.com>