Use domain-appropriate relation for `phrenic nerve`.
The term `phrenic nerve` was said to _supply_ the diaphragm. However the
`supplies` relation (RO:0002178) only applies to arteries, according
both to its definition (“relation between an arterial structure and
another structure…”) and its domain restriction (`artery`,
UBERON:0001637).

This commit replaces the `supplies` relation with the `innervates`
(RO:0002134), which was probably what was originally intended.