Replace 'supplies' with 'innervates' where appropriate.
The 'supplies' relation is only intended to be applied to arteries and
has a domain restriction to enforce that. Using it to indicate that a
nerve provides innervation to a muscle is wrong.

This PR fixes two terms ('inferior branch of oculomotor nerve' and
'superior branch of oculomotor nerve') that were using the 'supplies'
relation despite being nerves instead of arteries.

closes #2817