Remove clearly incorrect uses of 'contains process'.
The 'contains process' relation is defined as "a relation between an
independent continuant and a process, in which the process takes place
entirely within the independent continuant".

By this definition, the relation should not be used to merely indicate
that a process _may_ happen in a particular anatomical entity, if that
process can also take place in other, unrelated anatomical entities.

There are a handful of clearly incorrect uses of this relation in
Uberon, this commit removes them.

related to #2816.