Fix classification of 'cardiac muscle tissue'.
Re-classify 'cardiac muscle tissue' as part_of some 'primary circulatory
organ' rather than 'heart', as the latter is vertebrate-specific.

This aligns the Uberon term with the CL term 'cardiac muscle cell',
which is used in the definition of the Uberon term and is clearly
intended, in CL, to apply to a broader range of species than just the
vertebrates.