Remove 'tmp/core.owl' intermediate product.
The 'core.owl' product is no longer needed now. Its only purpose was to
be merged with phenoscape-ext, which is now part of Uberon itself.

The product was referred to in the 'quick_ac' rule (not called from
anywhere, but which may be called manually by an user); in this instance
we replace it by the $OWLSRC product.