Remove old rule about ENVO import.
The rule to create envo_import.owl had been overriden at some point
because ENVO was "broken". Don't know if ENVO is still broken, but in
any case we are no longer creating an envo_import.owl module since we
have switched to the base merging mechanism, so the override is no
longer needed.