Re-organise and re-comment the section on imports.
No logical change here. This commit simply re-organise the sections on
mirrors and imports to have a clearer flow:

* first all mirrors (both the standard, ODK-declared mirroring rules
  that we override, and the Uberon-specific custom mirrors);
* then the rules to make the "local-" imports: first the generic rule,
  then the handful of ontology-specific rules.

Comments are also re-written to match what is actually happening and to
call things accurately (instead of seeding confusion by calling
"mirrors" what are actually imports or the other way round).