Regroup all report-generating rules in the REPORTS section.
Rules that generate reports are scattered throughout the Makefile, so we
regroup them in the REPORTS section and organise them in somewhat
sensible subsections. Whenever possible, we add a brief comment to
explain what the report is about.

This does not concern all the reports about the bridges and the rest of
the composite-* stuff; those rules probably deserve their own section.