Remove the bridge to EMAP.
According to Chris Mungall, EMAP is dead. The bridge is currently
imported into collected-mouse-with-stages.owl, which should probably be
removed.

For now, we just remove the rules to generate the EMAP bridge. The
bridge file itself is kept as it is, so we can still build
collected-mouse-with-stages.owl until we decide whether we can remove it
as well or not.