Regroup subset-generating rules in the SUBSETS & VIEWS section.
Similar to the recent changes, we regroup all rules that create subsets
(and "views", which are merely subsets by another name) in a single
section.

Most rules are kept as they are, this commit is mostly just moving stuff
around. In particular, most calls to Owltools are left untouched, since
they are making use of features that are not available or not easily
replicated with ROBOT.