Fixing the bridge checks (part 1).
There are currently 4 "flavours" of bridge checks, with several
problems:

1. They all rely on Owltools.
2. The "quick" flavour, as implemented, does not do what the associated
   comments say it does.
3. There is actually no difference between the "standard" bridge check
   and the "full" bridge check.
4. The pre-requisites listed for each check do not necessarily
   correspond to the files that are actually needed.

Here we fix all problems by re-defining only three flavours of checks,
that differ by what is merged before running the reasoner:

* quick check: uberon + external-disjoints + bridge to external
  ontology;
* standard check: same as quick check + external ontology
* extra full check: same as standard check + pending-disjoints.

All three checks are implemented using ROBOT and with proper
prerequisites.

The top-level comment explaining the various flavours of checks is
removed in favour of the comments located just above the actual
implementation of the checks (this should help avoid another deviation
between the implementation and the comment in the future).

closes #3022