Fixing the BFO bridge checks.
(Part 3 of fixing the bridge checks.) The bridge to BFO cannot be
checked using the generic rules used for the other bridges, notably
because when the check is run (which typically is done under MIR=false
conditions), we do not have a local RO or local BFO mirror to use, so we
need to download the external ontology from its online location.

We add a comment to explain that and re-write the rules to use ROBOT
instead of Owltools, as for the other bridge checks.