Simplify the composite stages section.
The CSTAGES variable is no longer needed as it only expands to one file
(tmp/developmental-stage-ontologies/src/ssso-merged-uberon.obo) -- it is
probably a remnant of a time where the developmental stage ontologies
repository was not already providing a pre-built ssso-merged file.

Replace Owltools by ROBOT as all that's needed in this section is just
merging several files together.