Remove FBbt-nd.
FBbt-nd is a version of FBbt without disjointness axioms. This is not
needed as the imports/local-fbbt.owl mirror is already generated by,
among other operations, stripping disjointness axioms from the upstream
ontology.

FBbt-nd is only used to build composite-drosophila, which we update here
to use the imports/local-fbbt.owl mirror instead.