Make the composite-* stuff explicitly depend on update-stages.
We need the `ssso-merged-uberon.obo` file from the developmental stage
ontologies repo in order to build composite-metazoan (the file is
imported in collected-metazoan), so we better explicitly add the
update-stages step as a dependency.

Otherwise we only get that file as a side-effect of the declaration of
the emapa mirror.