Move the rule to generate external-disjoints.owl.
The rule to generate tmp/external-disjoints.owl is not a "Nico hack", it
is a prerequisite to all the bridge checks, so we move it to the BRIDGE
CHECKS section.

Of note, since the external disjointness axioms are maintained in a OBO
file (from which the .owl file is derived with the rule we are talking
about here), we could get rid of that rule and amend the bridge checks
to use the OBO version directly.