Generalize the use of external SSSOM set.
For now, FBbt is the only foreign ontology to provide its mappings as a
SSSOM set, but that may change in the future if SSSOM gets used more
widely.

So we generalise the rules that download the FBbt mapping set so that
they can be used to download any externally provided mapping set.

Likewise, the rule that generates the mappings.owl components (made of
cross-references derived from the FBbt SSSOM set) is generalised so that
it is ready to create a component from *all* externally provided mapping
sets, not only the FBbt one.