Remove unknown empty bridges.
The Uberon bridge to BSD and the CL bridge to "NIF Gross Anatomy" are
both empty and their intended source of truth is unknown (they are *not*
set up to be generated from cross-references, as most other bridges
are).

Empty and no source to fill them: they are useless and can be removed.