Set the edit file as SoT for the BFO bridge.
We have a bridge to BFO whose source of truth is currently unknown. We
align it with the other bridges by creating cross-references in the
Uberon edit file, so that the edit file becomes the source of truth and
the bridge can be (re-)generated in the same way as all the other
bridges.

Of note, several of the axioms in the bridge are redundant with axioms
already explicitly stated in Uberon, so we don't add cross-references
for those.