Remove the VAO bridge.
It seems VAO and VSAO are the same thing. VAO is listed as an
alternative prefix for VSAO in the bioregistry, and seemingly all
cross-references to VSAO terms in the Uberon edit file have a VAO
equivalent in the VAO bridge.

Since 1) we do not know where the VAO bridge comes from and 2) we have
already decided to remove the VSAO bridge because VSAO is obsolete,
there is no reason to keep this bridge.