Generate bridge files from SSSOM sets.
Dear reader, we have a lot to unpack here.

Very simply put, this commit replaces the old Perl script that generated
the bridge files from cross-references in Uberon and CL. Critically, it
does so by clearly separating the process in two independent steps:

* extracting the cross-references and make a SSSOM set out of them;
* generating the bridges out of the SSSOM set.

This allows for future flexibility as both steps can be modified
independently of the other, as long as the first step still results in a
SSSOM set and the second set still takes a SSSOM set as input.

(In particular, in the future we may switch to maintaining the mappings
directly in SSSOM, instead of using cross-references in the -edit file.
This would obviously make the first step above obsolete, but the second
step would be unaffected.)

Now for the details:

The first step is entirely done by the 'xref-extract' command of the
SSSOM plugin for ROBOT. We merge Uberon and CL (since we also need to
extract xrefs from CL, not only Uberon), then use xref-extract to
produce a SSSOM set. The command uses the 'treat-xrefs-as-...'
annotations found in the ontology to decide: 1) which cross-references
to extract (only the cross-references using a prefix declared in a
'treat-xrefs-as-...' annotations are considered), and 2) which mapping
predicate to use according to the following table:

* 'treat-xrefs-as-equivalent' -> skos:exactMatch
* 'treat-xrefs-as-has-subclass' -> skos:narrowMatch
* 'treat-xrefs-as-is_a' -> skos:broadMatch
* 'treat-xrefs-as-reverse-genus-differentia' -> semapv:crossSpeciesExactMatch

We use the same process to extract cross-references from ZFA, since ZFA
is the source of truth for the CL-to-ZFA mappings (but not for the
Uberon-to-ZFA mappings, for some reason).

In parallel, we fetch the SSSOM sets that are maintained by foreign
ontologies (for now, only FBbt), so that we end up with mapping sets
covering all the ontologies we want to bridge with.

The second step is done by the 'sssom-inject' command of the same SSSOM
plugin for ROBOT. Beyond the mapping sets themselves, that command
requires two additional files:

1) The bridge/bridges.rules file dictates which bridging axioms should
be generated for each mapping, depending on the mapping predicate and
the subject or object of the mapping. It also takes care of making sure
the mappings are in the expected orientation (Uberon/CL terms on the
object side), filtering mappings to anything else than Uberon/CL (in
case the externally provided mapping sets contain more than what we
need), and checking that all mappings concern classes that do exist in
Uberon/CL and are not obsolete. That file is written in the adhoc
"SSSOM/Transform-OWL" language, the documentation of which can be found
on the website of the SSSOM plugin.

2) The bridge/bridges.dispatch dictates where to write the bridging
axioms produced by the bridges.rules file. It does so using a tagging
system: each rule in bridges.rules is tagged to indicate the bridge to
which the axioms generated by that rule belong. The dispatch file
associates each tag to an output file and in addition allows to specify
some metadata for each output file.