Maintain the ruleset with M4 macros.
Since the bridge-generating rules are highly repetitive, they can easily
be automatically generated by some macros. This keeps the editor-facing
bridge/bridges.rules.m4 file easier to read and modify (for example, to
add or remove a bridge, an editor has simply to add or remove a single
'BRIDGE(...)' line).