Remove the tmp/seed.obo intermediate.
The seed.obo file was previously used as an intermediate to generate the
bridge file. It is no longer needed for this.

It was also used to generate the life-cycle-xrefs report, but there is
no reason not to use the -edit file as source for that step. The only
thing we need to be careful about is that seed.obo was pruned of
obsolete terms, while the -edit file obviously is not. But it is more
efficient to filter out obsolete terms directly in the SPARQL query,
instead of:

* use ROBOT to generate seed.obo;
* use (another instance of) ROBOT to reason over seed.obo;
* use a Perl script to hack the resulting OBO file and prune obsolete
  terms;
* use ROBOT again to run the SPARQL query.