Allow the bridge generation pipeline to run under MIR=false.
To generate the bridge, we need a version of CL both to extract the
cross-references from it, and to filter out any mappings to an
inexistent or obsolete CL class. We also need a version of ZFA to
extract the cross-references from it.

This is problematic as the bridge generation pipeline is typically run
under MIR=false conditions, so we may not have local CL and ZFA mirrors
available (unless we are on an editor machine and the editor has
refreshed the mirrors prior to running the pipeline).

There are several possible solutions:

1) For the specific purpose of generating the bridges, always download
CL and ZFA from their online location, regardless of MIR (basically
bypassing the ODK mirroring system). This is what the previous version
of the pipeline was doing. This is not completely unreasonable as we are
not using CL or ZFA to inject any axioms into a final product here, we
are just extracting their cross-references (though one might argue that
those cross-references ultimately end up as axioms in products such as
composite-metazoan).

2) Make the pipeline dependent on the mirrors -- that is, skipping the
pipeline entirely under MIR=false. This would mean that the bridges
would only be re-generated when the pipeline is specifically run under
MIR=true -- notably, they would never be re-generated during the CI
checks, since those run under MIR=false.

3) Try to have it both ways: use a local mirror when available (either
because we are under MIR=true, or because the mirrors had been refreshed
previously), and fallback to download CL and ZFA from their online
locations otherwise.

This commit implements option 3. I don't particularly like it (it's a
ugly hack), but I would like input from other maintainers before
deciding between options 1 and 2 (or possibly another option that I
didn't see), so this will do for now.