Create a report of "duplicated" cross-references.
When generating the SSSOM set from the cross-references in the -edit
file (and in CL), ignore any cross-references to a foreign term if there
are more than one, and produce a report listing all such cases.

This requires version 0.4.2 of the SSSOM plugin, which is not released
yet -- but it will be before this branch is ready for prime time.