Really avoid bridging one subject to more than one object.
We are already ensuring, when extracting the cross-references from
Uberon and CL, that we detect, ignore and report any case where the same
foreign term is mapped to more than one Uberon/CL term.

But we could still get duplicate mappings when we add to the mix the
externally provided sets -- either because an external set already
contains duplicate, or because of an overlap between the mappings
extracted the cross-references and the external sets.

So we add another layer of defense during the bridge generation step,
where we instruct sssom-inject to drop any mapping with a cardinality of
*:n (any mapping where the subject is mapped to many objects).